placed utility functions in utilities.js
